cli: hide 'connect' from picodata --help
'connect' is deprecated and it side tracks the customer journey.

Imagine I run picodata for the first time, execute
'picodata run' and then 'picodata help' to look for ways to connect
to the running node. The most obvious choice from help is to
use 'picodata connect'. However, the next *correct* step
is to use picodata admin to create a user and grant this user
privileges. If I execute 'picodata connect' it actually succeeds,
but I get a crippled console without any access rights.

Reduce confusion by removing the command from the help list.