test: test_offline_replicaset stopped working some time ago
Unfortunately the test was bad from the start, and it didn't break when
we changed the behavior that it was checking for.

The good new is the test was checking for behavior which we didn't want
to have: not being able to add an instance to the cluster if the only
replicaset in the cluster is offline.

At some point we fixed that and now we can add instances whenever as
long as there's raft quorum. But we don't know in which commit this
behavior changed.

This commit just fixes the test so it's no longer flaky.