ci: use locally builded tarantool for tarantool-module ci
When the tarantool module was in a separate repository,
the tarantool version was specified in the CI. The core/tarantool
release docker image was used. After integrating it
into the core/picodata repository, the test tarantool
version must match the tarantool-sys submodule version.

For testing, the tarantool binary is now built when creating
a basic build-base-image-fork image.
The TARANTOOL_VERSION variable in the CI tarantool/.gitlab-ci.yml
file is no longer needed.