ci: fix tarantool-module-trigger rules
CI job tarantool-module-trigger did not run in some cases when it should have.
The reason was that the MAIN_BRANCH variable was not available
in the context of this job. rules:inherit:variables was set to false.
It turned out that this affected not only the child pipeline,
but also the rules of the task itself.

To fix this, you need to include the MAIN_BRANCH variable
in the list of available variables in this task.