perf: use static linking for most tarantool symbols
Previously tarantool-module would define some number of symbols as
being linked dynamically at runtime with the help of `define_dlsym_reloc!`
macro.

The symbols defined that way would be looked up dynamically via `dlsym`
the first time the function was called and cached in a static variable
for future reuse, so there wouldn't be too much of a performance hit.

This was needed because tarantool-module was supposed to support
several different versions of tarantool which could have a different set
of exported symbols, so we wouldn't know at compile time if a given
symbol would be available or not, so we moved the decision to runtime.

This is good enough for most things but introduces an inherent
performance ceiling as it's not possible to optimize away the function
call because the function address is not known at compile time (or link
time).

This commit introduces a new feature `static_linking` for tarantool-module
which when enabled marks most of the external symbols as plain old
extern "C" symbols, which requires the compiler to link them at build
time. The new macro `define_extern_of_dlsym_reloc!` will conditionally
unwrap into `extern "C"` or `define_dlsym_reloc!` based on this feature
flag.

We enable this flag when importing tarantool-module into picodata and
therefore can take advantage of more advanced link-time optimizations.