feat(sql): initial anonymous block support
This commit introduces basic anonymous blocks.

An anonymous block is a sequence of statements that execute queries
transactionally.
Blocks are currently single-bucket, meaning that all the queries
within the block must be executed on the same bucket
(or have distribution any).

The following statements are implemented:
- QUERY: execute the given query;
- RETURN QUERY: execute the given query and return its result.

LET statements can be parsed, but always lead to
an error that they are not supported yet.

Example of a block containing RETURN QUERY and QUERY statements:
```sql
DO $$ BEGIN
  RETURN QUERY SELECT a FROM t WHERE pk = 1;
  UPDATE t SET a = a + 1 WHERE pk = 1;
END $$;
```