perf(sql): don't clone values during local values materialization
This commit avoids cloning of constants during local values
materialization. With this change we can benefit if there are many
parameters allocated on the heap. However, we don't get it for free -
to take values from the IR now we need to traverse it twice so we
exchange cpu cycles for memory here.