test: use tee -a so that logs don't get removed during self-exec
As part of bootstrap procedure picodata does a self-exec restart. During
this time it will spawn another logger process (if such option is
specified). We use `tee` in tests which by default overwrites the output
file if it already exists. Now we add a `-a` flag to `tee` so that the
output file is instead appended to.