feat(sql): support JSON_EXTRACT_PATH function
The JSON_EXTRACT_PATH function is used in SELECT queries
to extract data from a field in JSON format.

The first argument of the function is the JSON field from
which data is extracted. The second and subsequent arguments
of the function are path components specifying where
to extract data from the JSON.
They represent the results of expressions of type TEXT.