test(flaky): deflakify `test_vinyl_tmp_table`
Currently, secondary nodes may observe stale state unless the test
explicitly waits for the current vclock corresponding to the node that
executed the query.

This change adds the required synchronization to queries executed in
this test, to prevent incomplete data from being returned on secondary
nodes and eliminates the resulting test flakiness.

Closes: picodata#2126.