perf: don't do lua eval in Vclock::current
Before this change when getting current vclock we would do a lua eval
call, which involves starting up the lua parser, generating the lua byte
code and then executing that byte code, which is relatively expensive.

In some cases we're going to be calling Vclock::current rather
frequently, so it's important to avoid invoking the lua parser.

Now we use lua api to call the lua functions directly from rust, without
invoking the lua parser. This is still not ideal though as it would be
better to use a dedicated native API, but unfortunately we don't have
that yet.