feat: add tlua::Object::read function
Before this change the only way to get multiple values from lua was via
LuaRead implementation for rust tuples. For example if a lua function
returns multiple values you would call that function from rust and
specify the return type as `(T1, T2, ...)`.

This worked fine for types which don't capture the lua stack guard, such
as integers, floats, strings and aggregates consisting of them (array and structs).

But types which do capture the lua stack guard, for example LuaTable,
LuaFunction or StringInLua this doesn't work.

The details escape me at the moment, but I do vaguely remember that
adding support for that to impl LuaRead for (T1, ...) would be if not
impossible then very complicated to implement and would require a rather
substantial refactoring.

This commit introduces an alternative approach. If you need to get
multiple values from lua, you read the function result as an opaque
`tlua::Object` and then access the separate values via calls to
`Object::read`. So `object.read(0)` returns the first of the values,
`object.read(1)` returns the seconds etc.

This works for all LuaRead implementing types.