feat: instance will not become Online until it synchonizes with replicaset
Before this commit we had a couple of problems. Whenever an instance
was waking up (by setting it's target_state to Online) governor would
simply reconfigure the replicaset by adding the new instance to all
other instance's upstreams and after that the instance's current state
would be immediately change to Online.

The first problem with that is that the newly Online instance could
still be very out of date with the current master, so SQL queries routed
to that instance would output stale results.

Now the instance will only become Online once it's syncrhonized with
master, so stale probability of stale results is reduced.

The other problem is that this could lead to replication conflicts being
propagated to the whole replicaset. This was the cause of this issue
for example https://git.picodata.io/core/picodata/-/issues/2231. The
replicaset master switched over during DDL and as the result 2 replicas
applied the same DDL and that conflict got propagated.

Now if an instance has a replication conflict it will not be added to
other replica upstreams so the conflict will not be propagated.

As part the fix we have added a new column `sync_incarnation` to the
`_pico_instance` system table. Whenever the instance starts waking up
(or attempts to return after auto-offline by sentinel) it will increase
it's `target_state.incarnation` value (old behavior, no changes here),
which will automatically make the `sync_incarnation` be out of date, so
governor will know that this instance must be synchronized. After the
instance is synchronize `sync_incarnation` is updated to the same value
as `target_state.incarnation` (new behavior introduced in this commit).

Close #2231
Close #2235