fix: new governor_check_replication_error ALTER SYSTEM parameter
In a previous commit we added a new governor step on which it waits
until waking up instance synchronizes the replication. There's also
some code which checks if there is a replication error. Now we add code
which checks this replication error and makes the broken instance
automatically offline.

We also add an ALTER SYSTEM parameter governor_check_replication_error
which enables this feature. If it's disabled, this instance will likely
keep trying to become online and governor will keep trying to
synchronize it's replication and no other lower priority action (e.g. DDL)
will be handled during this time.

The option is added mainly for testing but also in case there's a bug
somewhere and we need to disable that check in production.