test(rolling): raise explicit error for tests against zero patch versions
In our current workflow, versions that contain zero component (e.g.
the first version after minor release will contain a zero in PATCH) are
considered not valid versions, because they are used only to mark
beginning of a new release series, so there are no packages built for them.
This causes rolling upgrade tests to fail, because they rely on packages
for Rocky Linux to use in CI. Particular case that we encountered
includes new minor 25.5 being branched without 25.5.1 tag yet. So in
master we need to start new release series, 25.6 which implies tagging
25.6.0. So we have 25.6.0 which attempts to test against 25.5.0 because
there is no 25.5.1. The solution to that is to always tag a new release
immediately at branch point. So there is no such problem.

This commit adds explicit error for that case.

Co-Authored-By: Dmitry Rodionov <d.rodionov@picodata.io>