fix: don't try to change master in offline replicasets
There was a bug where governor would find any replicaset with
`target_master_name` != `current_master_name` and try to change the
master even if all instances in that replicaset were Offline. As a
result governor would sit in a loop trying to do that impossible action,
and any other action with lower priority would be ignored. So any other
instance trying to become Online would also be skipped.

This is now fixed and governor will ignore replicasets with
`target_master_name` != `current_master_name` if `target_master_name`
is Offline.

Note that if governor can choose a new `target_master_name` it would, so
it's only possible for `target_master_name` to be Offline if every
instance in that replicaset is Offline.