fix: instances would be made Offline immediately after a new raft entry if not entries for some time
There was a bug where sentinel would incorrectly make instances Offline
when trying to determine if they are lagging behind leader over raft.

Instead of checking for how long their applied index is lagging behind
leader's we were checking for how long their applied index is behind
the leader's at this moment in time.

So basically if there weren't any raft entries for 60 minutes but then a
new entry is applied sentinel sees that a follower hasn't applied any
entries for 60 minutes and is lagging behind our applied index (because
we just applied one entry) and would make that instance Offline.

This is fixed now. Sentinel now checks the applied index of the
follower, then it checks how long ago did we apply the **next entry after
that**.

And follower will only be made auto-offline if all of these are true:
- follower's applied index is lagging behind leader's
- follower's hasn't applied any entries for `governor_auto_offline_timeout` seconds (30 by default)
- at least `governor_auto_offline_timeout` seconds has passed since
  leader applied the entry one after that which follower has applied last

For the purposes of this we now keep track of history of times when we
applied recent raft log entries (up to `governor_auto_offline_timeout`
seconds in the past)