fix: picodata_plugin:::sql::query used to leak memory
Before this fix we had a bug, where when returning tuples from
`pico_ffi_sql_query` we would not decrease the tuple's reference counter,
but on the plugin's side we would increase the reference counter. As a
result every tuple returned from plugin SQL would be leaked.

We fix this by not increasing the refcount on the plugin's side.

Note that there's a potentially better approach. When tarantool returns
tuples for box_* API it avoids memory leaks by keeping a `last_tuple`
global variable which keeps track of the last returned tuple. So
whenever a tuple is returned via FFI API it always has a reference from
that `last_tuple`, but if nobody increases that tuple's refcount, it
will be freed whenever upon the next call to box_* API, before assigning
to `last_tuple` tarantool first unref's the old one (thus freeing it if
nobody else added a ref).

In the future we could add something like that to picodata's plugin SDK