feat: add unlogged table creation
Data written to unlogged tables is not written to the write-ahead log, hence
such tables are not replicated. They can be created using the `CREATE
UNLOGGED TABLE...` syntax. Note that globally-distributed tables and
tables using the `vinyl` engine can't be unlogged.

Unlogged tables are implemented as data-temporary tarantool spaces.

Table is considered unlogged if the array inside of a new `opts` field
of `_pico_table` contains a map `{"unlogged": [true]}`.

Unlogged tables are truncated on instances that become read-only.